Added selecting, pie progress, and style

+++ src/app/Home/home.ts
@@ -1,4 +1,4 @@
-import { Component, signal , inject} from '@angular/core';
+import { Component, signal , inject, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import {Task} from './task';
 import { Task as TaskItem } from './tasks';
@@ -12,16 +12,33 @@ import {ListSnackbar} from './list-snackbar';
 import {ListFilterToggle} from './filter';
 import {MatDialog, MatDialogModule} from '@angular/material/dialog';
 import {DeleteDialog} from './delete-dialog';
+import {ProgressPie} from './progress-pie';
 
 @Component({
   selector: 'home',
-  imports: [Task, FormsModule, 
+  imports: [Task, FormsModule, ProgressPie,
     ListFilterToggle, 
     MatDatepickerModule, MatFormFieldModule, MatIconModule, MatInputModule, MatDialogModule],
   templateUrl: './home.html',
   styles: ''
 })
-export class Home {
+export class Home implements OnInit {
+  allTasks!: TaskItem[];
+
+  progress!: number;
+
+  displayedTasks!: TaskItem[];
+
+  ngOnInit(): void {
+    let storedData = this.storageService.getItem('taskList');
+    this.allTasks = storedData ? JSON.parse(storedData) : [ {
+    name: "welcome!", date : new Date('2026-4-28'), done : true
+    }];
+
+    // completed task count
+    this.progress = this.allTasks.reduce((acc, task) => task.done ? acc + 1 : acc, 0);
+    this.displayedTasks = this.allTasks;
+  }
 
   filter: "all" | "active" | "completed" = "all";
 
@@ -37,6 +54,12 @@ export class Home {
 
   snackBarMessage = "";
 
+  selectedTasks: string[] = [];
+
+  // eventually save and retrieve from local storage
+  //storedData = this.storageService.getItem('taskList');
+  //progress = this.allTasks.reduce((acc, task) => task.done ? acc + 1 : acc, 0);
+
   // constructor(private snackBar: MatSnackBar) {}
 
   openSnackBar() {
@@ -51,24 +74,22 @@ export class Home {
   readonly dialog = inject(MatDialog);
 
   openDialog() {
-    const dialogRef = this.dialog.open(DeleteDialog, {
-      data : {message : "Are you sure you want to delete the selected tasks?"}
-    });
+    if (this.selectedTasks.length == 0) {
+      this.snackBarMessage = "Select Tasks to Delete!";
+      this.openSnackBar();
+    }
+    else {
+      const dialogRef = this.dialog.open(DeleteDialog, {
+        data : {message : "Are you sure you want to delete the selected tasks?"}
+      });
 
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.deleteSelected();
-      }
-    });
+      dialogRef.afterClosed().subscribe(result => {
+        if (result) {
+          this.deleteSelected();
+        }
+      });
+    };
   }
-  
-  // eventually save and retrieve from local storage
-  storedData = this.storageService.getItem('taskList');
-  allTasks: TaskItem[] = this.storedData ? JSON.parse(this.storedData) : [ {
-    name: "welcome!", date : new Date('2026-4-28'), done : true
-  }];
-
-  displayedTasks = this.allTasks;
 
   saveList() {
     this.storageService.setItem('taskList', JSON.stringify(this.allTasks));
@@ -79,7 +100,8 @@ export class Home {
   addTask() {
     let findTask = this.allTasks.find(t => t.name == this.newTask);
     if (!findTask && this.newTask != '') {
-      this.allTasks.push({name : this.newTask, date: this.newDate,done: false});
+      this.allTasks = [...this.allTasks,{name : this.newTask, date: this.newDate,done: false}];
+      this.allTasks.sort((a, b) => new Date(a.date).getTime() - new Date(b.date).getTime());
       this.newTask = '';
       this.saveList();
     }
@@ -95,6 +117,7 @@ export class Home {
 
     if(completedTask) {
       completedTask.done = true;
+      //this.progress += 1;
       // console.log(this.allTasks);
       this.snackBarMessage = "Task complete! Good Job!";
       this.openSnackBar();
@@ -108,8 +131,7 @@ export class Home {
   }
 
   filterTasks() {
-    console.log("called");
-    console.log(this.filter);
+    this.progress = this.allTasks.reduce((acc, task) => task.done ? acc + 1 : acc, 0);
     if (this.filter == 'active') {
       this.displayedTasks = this.allTasks.filter(task => !task.done);
     }
@@ -122,6 +144,23 @@ export class Home {
   }
 
   deleteSelected() {
-    
+    if (this.selectedTasks.length == 0) {
+      this.snackBarMessage = "Select Tasks to Delete!";
+      this.openSnackBar();
+    }
+
+    this.selectedTasks.forEach(task => {
+      this.markDeleted(task);
+    });
+
+    this.selectedTasks = [];
+  }
+
+  selectTask(name: string) {
+    this.selectedTasks = [...this.selectedTasks, name];
+  }
+
+  unselectTask(name:string) {
+    this.selectedTasks = this.selectedTasks.filter(task => task != name);
   }
 }
